Rename module elasticjob-error-handler-impl to elasticjob-error-handler-type (#1538)
