+++ elasticjob-error-handler/elasticjob-error-handler-type/elasticjob-error-handler-email/pom.xml
@@ -22,14 +22,41 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>org.apache.shardingsphere.elasticjob</groupId>
-        <artifactId>elasticjob</artifactId>
+        <artifactId>elasticjob-error-handler-type</artifactId>
         <version>3.0.0-beta-SNAPSHOT</version>
     </parent>
-    <artifactId>elasticjob-error-handler</artifactId>
-    <packaging>pom</packaging>
+    <artifactId>elasticjob-error-handler-email</artifactId>
     
-    <modules>
-        <module>elasticjob-error-handler-spi</module>
-        <module>elasticjob-error-handler-type</module>
-    </modules>
+    <dependencies>
+        <dependency>
+            <groupId>org.apache.shardingsphere.elasticjob</groupId>
+            <artifactId>elasticjob-error-handler-spi</artifactId>
+            <version>${project.parent.version}</version>
+        </dependency>
+        
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+        </dependency>
+        
+        <dependency>
+            <groupId>com.sun.mail</groupId>
+            <artifactId>javax.mail</artifactId>
+        </dependency>
+        
+        <dependency>
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+        </dependency>
+        
+        <dependency>
+            <groupId>org.mockito</groupId>
+            <artifactId>mockito-core</artifactId>
+        </dependency>
+        
+        <dependency>
+            <groupId>org.mockito</groupId>
+            <artifactId>mockito-inline</artifactId>
+        </dependency>
+    </dependencies>
 </project>
