+++ elasticjob-error-handler/elasticjob-error-handler-type/pom.xml
@@ -8,7 +8,7 @@
   ~ the License.  You may obtain a copy of the License at
   ~
   ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
+  ~  
   ~ Unless required by applicable law or agreed to in writing, software
   ~ distributed under the License is distributed on an "AS IS" BASIS,
   ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -22,14 +22,17 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>org.apache.shardingsphere.elasticjob</groupId>
-        <artifactId>elasticjob</artifactId>
+        <artifactId>elasticjob-error-handler</artifactId>
         <version>3.0.0-beta-SNAPSHOT</version>
     </parent>
-    <artifactId>elasticjob-error-handler</artifactId>
+    <artifactId>elasticjob-error-handler-type</artifactId>
     <packaging>pom</packaging>
     
     <modules>
-        <module>elasticjob-error-handler-spi</module>
-        <module>elasticjob-error-handler-type</module>
+        <module>elasticjob-error-handler-general</module>
+        <module>elasticjob-error-handler-email</module>
+        <module>elasticjob-error-handler-sms</module>
+        <module>elasticjob-error-handler-wechat</module>
+        <module>elasticjob-error-handler-dingtalk</module>
     </modules>
 </project>
